Keep workspace clippy-clean outside vendor code
This change removes all non-vendor clippy warnings across the workspace while preserving existing behavior. Fixes include small idiomatic cleanups, targeted type aliases for cache tuples, async test-locking in RPC integration tests, and narrow lint allowances only where helper ergonomics justified them.

Constraint: vendor/** is treated as reference-only for warning cleanup

Constraint: behavior-preserving cleanup only; no feature or protocol changes

Rejected: broad refactor to async-fn trait signatures | higher behavioral and API churn risk

Rejected: suppress all clippy lints globally | would hide actionable future warnings

Confidence: high

Scope-risk: moderate

Directive: Keep vendor-origin warnings excluded and maintain non-vendor clippy zero-warning baseline

Tested: nix develop --command cargo clippy --workspace --all-targets --message-format short (0 non-vendor warnings); nix develop --command cargo build; nix develop --command cargo test

Not-tested: CI matrix variants beyond local nix shell

+++ crates/app/execute/evm/state/src/block_storage.rs
@@ -240,7 +240,7 @@ impl BlockStorage for RethStateDb {
 
             receipts.push(AppReceipt {
                 status: reth_receipt.success.into(),
-                cumulative_gas_used: reth_receipt.cumulative_gas_used as u64,
+                cumulative_gas_used: reth_receipt.cumulative_gas_used,
                 logs: reth_receipt.logs,
             });
         }
